+++ core/server/api_container/server/service_network/user_service_launcher/user_service_launcher.go
@@ -11,10 +11,9 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
-	"github.com/kurtosis-tech/free-ip-addr-tracker-lib/lib"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/user_service_registration"
 	"github.com/kurtosis-tech/kurtosis-core/server/api_container/server/service_network/user_service_launcher/files_artifact_expander"
 	"github.com/kurtosis-tech/stacktrace"
-	"net"
 )
 
 /*
@@ -23,11 +22,10 @@ Convenience struct whose only purpose is launching user services
 type UserServiceLauncher struct {
 	kurtosisBackend          backend_interface.KurtosisBackend
 	filesArtifactExpander    *files_artifact_expander.FilesArtifactExpander
-	freeIpAddrTracker        *lib.FreeIpAddrTracker
 }
 
-func NewUserServiceLauncher(kurtosisBackend backend_interface.KurtosisBackend, filesArtifactExpander *files_artifact_expander.FilesArtifactExpander, freeIpAddrTracker *lib.FreeIpAddrTracker) *UserServiceLauncher {
-	return &UserServiceLauncher{kurtosisBackend: kurtosisBackend, filesArtifactExpander: filesArtifactExpander, freeIpAddrTracker: freeIpAddrTracker}
+func NewUserServiceLauncher(kurtosisBackend backend_interface.KurtosisBackend, filesArtifactExpander *files_artifact_expander.FilesArtifactExpander) *UserServiceLauncher {
+	return &UserServiceLauncher{kurtosisBackend: kurtosisBackend, filesArtifactExpander: filesArtifactExpander}
 }
 
 /**
@@ -39,10 +37,8 @@ Returns:
 */
 func (launcher UserServiceLauncher) Launch(
 	ctx context.Context,
-	serviceGUID service.ServiceGUID,
-	serviceId service.ServiceID,
+	registrationGuid user_service_registration.UserServiceRegistrationGUID,
 	enclaveId enclave.EnclaveID,
-	ipAddr net.IP,
 	imageName string,
 	privatePorts map[string]*port_spec.PortSpec,
 	entrypointArgs []string,
@@ -63,7 +59,7 @@ func (launcher UserServiceLauncher) Launch(
 	// First expand the files artifacts into volumes, so that any errors get caught early
 	// NOTE: if users don't need to investigate the volume contents, we could keep track of the volumes we create
 	//  and delete them at the end of the test to keep things cleaner
-	artifactUuidsToVolumes, err := launcher.filesArtifactExpander.ExpandArtifactsIntoVolumes(ctx, serviceGUID, usedArtifactUuidSet)
+	artifactUuidsToVolumes, err := launcher.filesArtifactExpander.ExpandArtifactsIntoVolumes(ctx, registrationGuid, usedArtifactUuidSet)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred expanding the requested files artifacts into volumes")
 	}
@@ -84,11 +80,9 @@ func (launcher UserServiceLauncher) Launch(
 
 	launchedUserService, err := launcher.kurtosisBackend.CreateUserService(
 		ctx,
-		serviceId,
-		serviceGUID,
+		registrationGuid,
 		imageName,
 		enclaveId,
-		ipAddr,
 		privatePorts,
 		entrypointArgs,
 		cmdArgs,